add Liberapay: field with username as data
Liberapay was originally included using a numeric ID, since they had
not yet finalized the public URLs.  Now it is a username.  So this
logic prefers the username in Liberapay: field, and keeps the old
LiberapayID: to ease migration.  LiberapayID: will not override
Liberapay:.  Clients are expected to prefer Liberapay: over LiberapayID:

# Conflicts:
#	fdroidserver/update.py
#	tests/metadata/info.guardianproject.checkey.yml
#	tests/metadata/obb.main.oldversion.yml
#	tests/metadata/org.adaway.yml
#	tests/update.TestCase