scanner: should not exit with error when apkanalyzer fails
apkanalyzer produces useful output when it can run, but it does not
support all recent JDK versions, and also some DEX versions, so this
cannot count on it to always produce useful output or even to run
without exiting with an error.

211dd65ff0263e3b680dcd21f9e0efa3dd7e50d7 was based on false
assumptions that apkanalyzer can always produce output.

fdroiddata!8585
fdroiddata!8584