set up config['ndk_paths'] on demand by reading source.properties
ndk_paths is set up right after installing the NDK and right before
ndk_paths is used in Build.ndk_path().

A proper NDK install always has a _source.properties_ which declares
the revision and release strings. That is used as the source of
revision and release.  This only reads from the local filesystem, and
reads the official source of version info (_source.properties_),
thereby avoiding any potential network calls that sdkmanager might
do. This implementation was inspired by @uniqx's:

https://gitlab.com/uniqx/fdroidserver/-/commit/ed931d47d728291ba6b428e0fcf7b53e52476d98#27b997d583cd1c0a05b2213437e0d9fe6536bfe3