set up new mechanism for unlisted subcommands
For the build automation, there will be lots of subcommands that are meant
to run in scripted environments.  They are not intended for interactive use.
Although they sometimes might be useful interactively, they might also
sometimes be dangerous to run interactively because they rely on the
presense of a VM/container, modify the local environment, or even run
things as root.