'manifest' in internal struct; remove single use of namedtuple
Everywhere else, just plain data structures are used, like dict, list, etc.
Plus even the code that accesses these namedtuples is not using the features
of the namedtuple.

These are steps towards refactoring the code so that index-v2 is the
internal data structure of fdroid update.  Then ultimately, we'll be
able to remove the whole convert_version() function and streamline
other parts of the code.  Then the data structure conversions will
only need to happen in the code that generates the old index files.